+++ apps/client-website/src/app/services/modal-reference/index.ts
@@ -1,0 +1,1 @@
+export * from './modal-reference.service';
